+++ E-Commerce.Application/Features/Coupons/Commands/UpdateCoupon/UpdateCouponCommand.cs
@@ -3,7 +3,7 @@
     public record UpdateCouponCommand : IRequest<CouponDto>
     {
         public string guid { get; init; } = string.Empty;
-        public decimal DiscountValue { get; init; }
+        public double DiscountValue { get; init; }
         public DiscountType DiscountType { get; init; }
         public DateTimeOffset StartDate { get; init; }
         public DateTimeOffset EndDate { get; init; }
